separate runm-metadata functionality
To make the prototyping easier, I am just using a couple tables in the
same database as the resource data models to store what will be in the
runm-metadata service's etcd3 data store.

This patch just pulls functions that were getting and setting
object_types and object_names table records into its own module for
simplicity and clean separation from the resource database models.
